developement_r - simpel logs

+++ utils/src/SocketUtils.cpp
@@ -20,7 +20,7 @@ void sendMsg(int socket, const std::string& msg) {
 void sendXBytes(int socket, unsigned int x, void *buffer) {
     char* buffer_ = (char*)buffer;
     int sendBytes = 0;
-    while (sendBytes < x) {
+    while (sendBytes < (int)x) {
         int packageSize = send(socket, buffer_ + sendBytes, x - sendBytes, 0);
         if (packageSize <= 0)
             std::cerr << "couldn't send bytes when expected";
@@ -48,7 +48,7 @@ std::string readMsg(int socket) {
 void readXBytes(int socket, unsigned int x, void *buffer) {
     char* buffer_ = (char*)buffer;
     int bytesRead = 0;
-    while (bytesRead < x)
+    while (bytesRead < (int)x)
     {
         int result; result = read(socket, buffer_ + bytesRead, x - bytesRead);
         if (result < 1) {
